Add features to manage server members and invitations
This commit adds new features to manage server members and invitations. The ability to invite, kick, or change roles of members has been implemented. Two new modals were also added for inviting people to a server and editing server details, along with a new hook to get the origin of the URL for generating invite links. The server header now allows opening of the modal to manage members of a server as well.